Fix ambiguous calls to math functions in VC++ headers

+++ item.cpp
@@ -2157,7 +2157,7 @@ bool item::reload(player &u, char ammo_invlet)
   }
   else if (reload_target->typeId() == "adv_UPS_off" || reload_target->typeId() == "adv_UPS_on") {
       int charges_per_plut = 500;
-      int max_plut = std::floor( (max_load - reload_target->charges) / charges_per_plut );
+      int max_plut = std::floor( static_cast<float>((max_load - reload_target->charges) / charges_per_plut) );
       int charges_used = std::min(ammo_to_use->charges, max_plut);
       reload_target->charges += (charges_used * charges_per_plut);
       ammo_to_use->charges -= charges_used;
